+++ portal/src/pages/Blog/Blog.const.ts
@@ -1,0 +1,12 @@
+export const BLOG_POSTS = [
+  {
+    path: '/whats-new',
+    title: 'Bear 1.3.2',
+    date: '2026-08-28',
+  },
+  {
+    path: '/changelog',
+    title: 'Changelog',
+    date: '2026-08-28',
+  },
+] as const;