+++ portal/src/pages/DesignKits/DesignKits.const.ts
@@ -1,0 +1,3 @@
+import { GITHUB_URL } from '@/constants/navigation.const';
+
+export const DESIGN_KITS_GITHUB = GITHUB_URL;